+++ frontend/src/config/apiConfig.js
@@ -1,0 +1,5 @@
+// Transformamos y validamos las variables
+export const config = {
+  // Url de la api
+  apiUrl: import.meta.env.VITE_API_URL,
+};